gestion_livres

+++ Backend/controllers/books.js
@@ -1,29 +1,90 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 
 // Logique pour chaque fonction
 
 exports.createBook = (req, res, next) => {
-  delete req.body._id;
-  const book = new Book ({
-    ...req.body   
-  });
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete  bookObject._userId;
+    const book = new Book({
+      ...bookObject,
+      userId: req.auth.userId,
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    });
+
   book.save()
-    .then(() => res.status(201).json({message: 'Objet enregistré!'}))
+    .then(() => res.status(201).json({
+      message: 'Objet enregistré!'}))
     .catch(error => res.status(400).json({ error: error}));
 }
 
 exports.modifyBook = (req, res, next) => {    
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.setMaxListeners(200).json({ message: 'Objet modifié !'}))
-    .catch(error => res.status(400).json({ error }));
+  const bookObject = req.file ? {
+    ...JSON.parse(req.body.book),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+
+  delete bookObject._userId;
+  Book.findOne({_id: req.params.id})
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(403).json({ message: 'Non-autorisé'})
+      } else {
+        Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
+        .then(() => res.status(200).json({message: 'Objet modifié!'}))
+        .catch(error => res.status(401).json({error}));
+      }
+    });
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
+  Book.findOne({_id: req.params.id})
+    .then(book => {
+      if (book.userId != req.auth.userId) {
+        res.status(403).json({message: 'Non-autorisé'});
+      } else {
+        const filename = book.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({_id: req.params.id})
+            .then(() => { res.status(200).json({message: 'Objet supprimé !'}); })
+            .catch(error => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch(error => res.status(500).json({ error }));
 }
+
+exports.rateBook = (req, res, next) => {
+  const bookId = req.params.id;
+  const userId = req.auth.userId;
+  const grade = req.body.rating;
+
+  if (grade < 0 || grade > 5) {
+    return res.status(400).json({ message: 'Note invalide. Elle doit être entre 0 et 5.' });
+  }
+
+  Book.findOne({ _id: bookId })
+    .then(book => {
+      if (!book) {
+        return res.status(404).json({ message: 'Livre non trouvé.' });
+      }
+
+      // Ajouter la nouvelle note
+      book.ratings.push({ userId, grade });
+
+      // Recalculer la moyenne
+      const sum = book.ratings.reduce((acc, r) => acc + r.grade, 0);
+      book.averageRating = (sum / book.ratings.length).toFixed(2);
+
+      // Sauvegarder
+      book.save()
+        .then(updatedBook => res.status(200).json(updatedBook))
+        .catch(error => res.status(500).json({ error }));
+    })
+    .catch(error => res.status(500).json({ error }));
+};
 
 exports.getOneBook = (req, res, next) => {  
   Book.findOne({ _id: req.params.id })
@@ -36,3 +97,11 @@ exports.getAllBook = (req, res, next) => {
   .then(books => res.status(200).json(books))
   .catch(error => res.status(400).json({error}))
 }
+
+exports.getBestRatedBooks = (req, res, next) => {
+  Book.find()
+    .sort({ averageRating: -1 }) // tri décroissant
+    .limit(3)
+    .then(books => res.status(200).json(books))
+    .catch(error => res.status(400).json({ error }));
+};
